Should Inspect => Console Tab
Should do this to inspect why code does not run as expected:
Step 1: Click on the arrow on the far left of the Console Tab
Step 2: Click on the HTML element, in this case, the Button to start inspecting it
Step 3: Go to EventListerners Tab to see if the button receive the Click. In this case, at first, when the code has a bugs, the EventListerners Tab is blank. 
After fixing the code, the EventListerners tab has "click"